+++ apps/backend/src/libs/modules/task-scheduler/libs/types/task-scheduler.type.ts
@@ -1,0 +1,7 @@
+type TaskFunction = () => void;
+
+type TaskScheduler = {
+	start: (schedule: string, task: TaskFunction) => void;
+};
+
+export { type TaskScheduler };