+++ apps/backend/src/modules/notifications/libs/types/types.ts
@@ -1,0 +1,7 @@
+export {
+	type NotificationBulkCreateRequestDto,
+	type NotificationBulkCreateResponseDto,
+	type NotificationCreateRequestDto,
+	type NotificationGetAllItemResponseDto,
+	type NotificationGetAllResponseDto,
+} from "@git-fit/shared";