+++ apps/frontend/src/pages/projects/projects.tsx
@@ -10,6 +10,7 @@ import { DataStatus, PermissionKey } from "~/libs/enums/enums.js";
 import { checkHasPermission } from "~/libs/helpers/helpers.js";
 import {
 	useAppDispatch,
+	useAppForm,
 	useAppSelector,
 	useCallback,
 	useEffect,
@@ -55,6 +56,11 @@ const Projects = (): JSX.Element => {
 		projectStatus,
 		projectsTotalCount,
 	} = useAppSelector(({ projects }) => projects);
+
+	const { control, errors } = useAppForm({
+		defaultValues: { search },
+		mode: "onChange",
+	});
 
 	const handleSearchChange = useCallback(
 		(value: string) => {
@@ -152,6 +158,7 @@ const Projects = (): JSX.Element => {
 		},
 		[handleDeleteConfirmationModalOpen],
 	);
+
 	const handleProjectCreateSubmit = useCallback(
 		(payload: ProjectCreateRequestDto) => {
 			void dispatch(projectActions.create(payload));
@@ -199,7 +206,13 @@ const Projects = (): JSX.Element => {
 					</div>
 				)}
 			</header>
-			<ProjectsSearch onChange={handleSearchChange} />
+			<ProjectsSearch
+				control={control}
+				errors={errors}
+				name="search"
+				onChange={handleSearchChange}
+			/>
+
 			{isLoading ? (
 				<Loader />
 			) : (