+++ packages/shared/src/modules/notifications/libs/types/notification-get-all-item-response-dto.type.ts
@@ -1,0 +1,7 @@
+type NotificationGetAllItemResponseDto = {
+	createdAt: string;
+	id: number;
+	payload: string;
+};
+
+export { type NotificationGetAllItemResponseDto };